make the StorageParticipant play nice with #flunk
on_cancel and gone workitem document issue

+++ lib/ruote/part/storage_participant.rb
@@ -114,6 +114,8 @@ module Ruote
 
       doc = fetch(fei)
 
+      return unless doc
+
       r = @context.storage.delete(doc)
 
       on_cancel(fei, flavour) if r != nil
